Parameter utils changes

+++ savu/plugins/basic_operations/arithmetic_operations_tools.py
@@ -11,7 +11,7 @@ class ArithmeticOperationsTools(PluginTools):
         """
         scalar_value:
               visibility: basic
-              dtype: int
+              dtype: float
               description: A scalar value value for arithmetic operation
                (if not in metadata).
               default: None
